electron-react-fastapi-todo/todo/pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.0"
description = ""
authors = ["unerue <unerue@me.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.97.0"}
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.3"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
